zmiana struktury projektu: podział na views i api, recreate_db przesunięte w odpowiednie miejsce, aktualizacja readme,
