+++ recreate_db.py
@@ -14,14 +14,13 @@ def create_user(username, password, firstname):
 
 
 if __name__ == '__main__':
-    script_path = os.path.realpath(sys.argv[0])
-    dir_path = os.path.dirname(script_path)
-    db = sql.connect(dir_path + "/db.sqlite")
+    PATH_TO_APP = "zeton"
+    db = sql.connect(PATH_TO_APP + "/db.sqlite")
 
-    with open('sql/01_db_init.sql') as f:
+    with open(PATH_TO_APP + '/sql/01_db_init.sql') as f:
         db.executescript(f.read())
 
-    with open('sql/02_insert_test_data.sql') as f:
+    with open(PATH_TO_APP + '/sql/02_insert_test_data.sql') as f:
         db.executescript(f.read())
 
     db.close()
