+++ run.py
@@ -2,6 +2,7 @@ import sys, os
 
 sys.path.append(os.getcwd())
 
-from zeton.main import app
+import zeton
 
+app = zeton.create_app()
 app.run()
