+++ zeton/static/css/main.css
@@ -31,11 +31,6 @@ body {
 
 }
 
-.logout-link a {
-    text-decoration: none;
-    font-size: large;
-    background: red;
-}
 
 #dodawanie {
